Crate bevy_ecs_macros

Macros

impl_param_set

Derive Macros

Bundle

Component

Event

QueryData

Implement QueryData to use a struct as a data parameter in a query

QueryFilter

Implement QueryFilter to use a struct as a filter parameter in a query

Resource

ScheduleLabel

Derive macro generating an impl of the trait ScheduleLabel.

States

SystemParam

Implement SystemParam to use a struct as a parameter in a system

SystemSet

Derive macro generating an impl of the trait SystemSet.