Crate bevy_ecs_macros

Macros

all_tuples

impl_param_set

Derive Macros

AmbiguitySetLabel

Generates an impl of the AmbiguitySetLabel trait.

Bundle

Component

RunCriteriaLabel

Generates an impl of the RunCriteriaLabel trait.

StageLabel

Generates an impl of the StageLabel trait.

SystemLabel

Generates an impl of the SystemLabel trait.

SystemParam

Implement SystemParam to use a struct as a parameter in a system

WorldQuery

Implement WorldQuery to use a struct as a parameter in a query