Crate bevy_ecs_macros

Macros for deriving ECS traits.

Derive Macros

Bundle

Implement the Bundle trait.

Component

Cheat sheet for derive syntax, see full explanation and examples on the Component trait doc.

EntityEvent

Cheat sheet for derive syntax, see full explanation on EntityEvent trait docs.

Event

Implement the Event trait.

FromWorld

Implement the FromWorld trait.

MapEntities

Implement the MapEntities trait.

Message

Implement the Message trait.

QueryData

Implement QueryData to use a struct as a data parameter in a query

QueryFilter

Implement QueryFilter to use a struct as a filter parameter in a query

Resource

Implement the Resource trait.

ScheduleLabel

Derive macro generating an impl of the trait ScheduleLabel.

SystemParam

Implement SystemParam to use a struct as a parameter in a system

SystemSet

Derive macro generating an impl of the trait SystemSet.